+++ tui/timer/init.go
@@ -1,0 +1,10 @@
+package timer
+
+import tea "github.com/charmbracelet/bubbletea"
+
+// INIT
+// ----
+
+func (m Model) Init() tea.Cmd {
+	return m.timer.Init()
+}
